refactor label generation
